[GREEN] Add AccountRecoveryForm and change account recovery page.
